change byte_length to size_t

+++ src/ppidd.cpp
@@ -937,7 +937,7 @@ extern "C" {
    Synchronously write to the file specified by the file handle.
    Writes number of bytes to the file identified by handle at location offset.
 \* ******************************************************************************************** */
-   int PPIDD_Eaf_write(int handle,double byte_offset,void *buff,int64_t byte_length) {
+   int PPIDD_Eaf_write(int handle,double byte_offset,void *buff,size_t byte_length) {
     switch (ppidd_impl) {
 #ifdef HAVE_MPI_H
 #ifdef HAVE_GA_H
@@ -958,7 +958,7 @@ extern "C" {
    Writes number of bytes to the file identified by handle at location offset.
    Operation is guaranteed to be complete when eaf_wait called with request_id argument returns.
 \* ******************************************************************************************** */
-   int PPIDD_Eaf_awrite(int handle,double byte_offset,void *buff,int64_t byte_length,int *request_id) {
+   int PPIDD_Eaf_awrite(int handle,double byte_offset,void *buff,size_t byte_length,int *request_id) {
     switch (ppidd_impl) {
 #ifdef HAVE_MPI_H
 #ifdef HAVE_GA_H
@@ -978,7 +978,7 @@ extern "C" {
    Synchronously read from the file specified by the file handle.
    Reads number of bytes to the file identified by handle at location offset.
 \* ******************************************************************************************** */
-   int PPIDD_Eaf_read(int handle,double byte_offset,void *buff,int64_t byte_length) {
+   int PPIDD_Eaf_read(int handle,double byte_offset,void *buff,size_t byte_length) {
     switch (ppidd_impl) {
 #ifdef HAVE_MPI_H
 #ifdef HAVE_GA_H
@@ -999,7 +999,7 @@ extern "C" {
    Reads number of bytes to the file identified by handle at location offset.
    Operation is guaranteed to be complete when eaf_wait called with request_id argument returns.
 \* ******************************************************************************************** */
-   int PPIDD_Eaf_aread(int handle,double byte_offset,void *buff,int64_t byte_length,int *request_id) {
+   int PPIDD_Eaf_aread(int handle,double byte_offset,void *buff,size_t byte_length,int *request_id) {
     switch (ppidd_impl) {
 #ifdef HAVE_MPI_H
 #ifdef HAVE_GA_H
